add local vimrc file content reconfig check

+++ install.sh
@@ -5,7 +5,7 @@ set -eo pipefail
 rootpath=/tmp/mrmgxxxx/vim/
 rm -rf $rootpath && mkdir -p $rootpath
 
-# install auto commands
+# install base commands
 run_yum_cmd=0
 command -v yum >/dev/null 2>&1 || run_yum_cmd=1
 if [ "$run_yum_cmd" -ne 1 ]; then
@@ -13,7 +13,7 @@ yum -y install gcc git wget make clang llvm the_silver_searcher >> $rootpath/ins
 else
 apt-get -y install gcc git wget make clang llvm silversearcher-ag >> $rootpath/install.log 2>&1
 fi
-echo -e "\e[34;1m🌈  Commands gcc git wget make clang llvm ag install successfully !\033[0m"
+echo -e "\e[34;1m🌈  Commands gcc/git/wget/make/clang/llvm/ag install successfully!\033[0m"
 
 # check local vim version
 need_install_vim=0
@@ -27,7 +27,7 @@ else
         echo -e "\e[34;1m🐱  Found local vim version $version which need to upgrade version to 8.0+ ...\033[0m"
         need_install_vim=1
     else
-        echo -e "\e[34;1m👀  Local vim version $version already installed !\033[0m"
+        echo -e "\e[34;1m👀  Local vim version $version is already installed!\033[0m"
     fi
 fi
 
@@ -45,32 +45,52 @@ if [ "$need_install_vim" -eq 1 ]; then
     version=`vim --version | head -n 1 | awk -F ' ' '{print $5}'`
     major=`echo $version | awk -F '.' '{print $1}'`
     if [ $major -lt 8 ]; then
-        echo -e "\e[34;1m😭  New vim version install failed !\033[0m"
+        echo -e "\e[34;1m😭  New vim version install failed!\033[0m"
         exit 1
     else
-        echo -e "\e[34;1m🌈  New vim version $version install successfully !\033[0m"
+        echo -e "\e[34;1m🌈  New vim version $version install successfully!\033[0m"
     fi
 fi
 
-# install and config vim-plug
+# check local vim-plug
 cd $rootpath
-rm -rf ~/.vim* && mkdir -p ~/.vim/autoload/
-wget -N https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim -P ~/.vim/autoload/ >> $rootpath/install.log 2>&1
-wget -N https://raw.githubusercontent.com/mrmgxxxx/vim/master/vimrc -O ~/.vimrc >> $rootpath/install.log 2>&1
+need_config_vim=0
+wget -N https://raw.githubusercontent.com/mrmgxxxx/vim/master/vimrc >> $rootpath/install.log 2>&1
+if [ -f "${HOME}/.vimrc" ]; then
+    diff $rootpath/vimrc ${HOME}/.vimrc >/dev/null 2>&1
+    if [ $? != 0 ]; then
+        echo -e "\e[34;1m😥  The ${HOME}/.vimrc file is not correct, reconfig vim-plug now ...\033[0m"
+        need_config_vim=1
+    else
+        echo -e "\e[34;1m👀  The vim-plug is already configed!\033[0m"
+    fi
+else
+    echo -e "\e[34;1m😥  Not found ${HOME}/.vimrc file, reconfig vim-plug now ...\033[0m"
+    need_config_vim=1
+fi
+
+# config vim-plug
+cd $rootpath
+if [ "$need_config_vim" -eq 1 ]; then
+    rm -rf ~/.vim* && mkdir -p ~/.vim/autoload/
+    wget -N https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim -P ~/.vim/autoload/ >> $rootpath/install.log 2>&1
+    cp -rf $rootpath/vimrc ${HOME}/.vimrc
+    echo -e "\e[34;1m🌈  Install and config vim-plug successfully!\033[0m"
+fi
 
 # install fzf
 need_install_fzf=0
 command -v fzf >/dev/null 2>&1 || need_install_fzf=1
 if [ "$need_install_fzf" -eq 1 ]; then
-    echo -e "\e[34;1m😥  Not found fzf command, and need to install now ...\033[0m"
+    echo -e "\e[34;1m😥  Not found fzf command, install now ...\033[0m"
     cd $rootpath
     rm -rf ~/.fzf*
     git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf >> $rootpath/install.log 2>&1
     ~/.fzf/install --all >> $rootpath/install.log 2>&1
     version=`fzf --version | awk -F ' ' '{print $1}'`
-    echo -e "\e[34;1m🌈  Command fzf $version install successfully !\033[0m"
+    echo -e "\e[34;1m🌈  Command fzf $version install successfully!\033[0m"
 else
-    echo -e "\e[34;1m👀  Local fzf command already installed !\033[0m"
+    echo -e "\e[34;1m👀  Local fzf command is already installed!\033[0m"
 fi
 
 echo -e "\e[34;1m\n🐸  Enjoy It ~ \n \033[0m"
